requirements.txt
anyio==4.8.0
click==8.1.8
graphql-core==3.2.6
h11==0.14.0
idna==3.10
libcst==1.6.0
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
strawberry-graphql==0.262.1
typer==0.15.2
typing_extensions==4.12.2
uvicorn==0.34.0
